requirements.txt
Django==4.1.2
django-ninja==0.19.1
orjson-3.8.2